+++ .travis.yml
@@ -4,6 +4,7 @@ sudo: required
 language: go
 
 go:
+  - 1.7.x
   - 1.8.x
   - tip
