+++ plugin/plugin.go
@@ -1,12 +1,10 @@
-package containerd
+package plugin
 
 import (
 	"fmt"
-	"path/filepath"
-	"plugin"
-	"runtime"
 	"sync"
 
+	"github.com/docker/containerd"
 	"github.com/docker/containerd/content"
 
 	"golang.org/x/net/context"
@@ -29,7 +27,7 @@ type Registration struct {
 type InitContext struct {
 	Root     string
 	State    string
-	Runtimes map[string]Runtime
+	Runtimes map[string]containerd.Runtime
 	Store    *content.Store
 	Config   interface{}
 	Context  context.Context
@@ -72,40 +70,4 @@ func Register(name string, r *Registration) error {
 
 func Registrations() map[string]*Registration {
 	return register.r
-}
-
-// loadPlugins loads all plugins for the OS and Arch
-// that containerd is built for inside the provided path
-func loadPlugins(path string) error {
-	abs, err := filepath.Abs(path)
-	if err != nil {
-		return err
-	}
-	pattern := filepath.Join(abs, fmt.Sprintf(
-		"*-%s-%s.%s",
-		runtime.GOOS,
-		runtime.GOARCH,
-		getLibExt(),
-	))
-	libs, err := filepath.Glob(pattern)
-	if err != nil {
-		return err
-	}
-	for _, lib := range libs {
-		if _, err := plugin.Open(lib); err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-// getLibExt returns a platform specific lib extension for
-// the platform that containerd is running on
-func getLibExt() string {
-	switch runtime.GOOS {
-	case "windows":
-		return "dll"
-	default:
-		return "so"
-	}
 }
