+++ plugin/plugin_other.go
@@ -1,0 +1,8 @@
+// +build !go1.8 windows
+
+package plugin
+
+func loadPlugins(path string) error {
+	// plugins not supported until 1.8
+	return nil
+}
